Dev & Infra/Dev/requirements.txt
Flask
Flask-CORS
Flask-JWT-Extended
mysql-connector-python
Werkzeug
requests
